fix: wrong chapter order, regression from a2daaf4
chapters will be in random order due to human_sort() being called on the whole index in convert_pages. the key for sorting is the absolute path of each page, which features the random UUID of each Archive.

learning time: random UUIDs are random.

only sort pages when extracting instead, since the culprit appears to be Window's ZLIB being non-standard (Windows moment)